+++ src/main/java/roomescape/domain/reservation/dto/request/ReservationCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package roomescape.dto.request;
+package roomescape.domain.reservation.dto.request;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
