+++ src/main/java/roomescape/domain/reservation/repository/InMemoryReservationRepository.java
@@ -1,11 +1,11 @@
-package roomescape.repository;
+package roomescape.domain.reservation.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.domain.Reservation;
 
 @Repository
 public class InMemoryReservationRepository implements ReservationRepository {
